Separate database role from repository layer
