+++ sephora-test/sephora-test/App/Repository/Repository.swift
@@ -10,10 +10,6 @@ import Foundation
 
 protocol RepositoryType: AnyObject {
     func getProductList(completion: @escaping (CompletionResult<Product>) -> Void, error: @escaping (String) -> Void)
-    
-    func saveProductItems(productItems: ProductItem)
-    func deleteAllListInDataBase()
-    func getProductItems(callback: @escaping ([ProductItem]) -> Void)
 }
 
 class Repository: RepositoryType {
@@ -44,35 +40,6 @@ class Repository: RepositoryType {
                 print("error = \(String(describing: error))")
             }
         }
-    }
-    
-    func saveProductItems(productItems: ProductItem) {
-        let productObject = ProductObject(context: context.stack.context)
-        productObject.idProduct = productItems.id
-        productObject.descriptionProduct = productItems.description
-        productObject.locationProduct = productItems.location
-        productObject.imageUrlProduct = productItems.image
-
-        context.stack.saveContext()
-    }
-
-    func deleteAllListInDataBase() {
-        let requestProduct: NSFetchRequest<ProductObject> = ProductObject.fetchRequest()
-        guard let productListItems = try? self.context.stack.context.fetch(requestProduct) else { return }
-        guard !productListItems.isEmpty else { return }
-        self.productObjects = productListItems
-        self.productObjects.enumerated().forEach { (_, object) in
-            self.context.stack.context.delete(object)
-            self.context.stack.saveContext()
-        }
-    }
-
-    func getProductItems(callback: @escaping ([ProductItem]) -> Void) {
-        let requestProduct: NSFetchRequest<ProductObject> = ProductObject.fetchRequest()
-        guard let productItems = try? context.stack.context.fetch(requestProduct) else { return }
-        self.productObjects = productItems
-        let product: [ProductItem] = productItems.map { return ProductItem(object: $0) }
-        callback(product)
     }
 }
